+++ components/card/camera-scan-view.tsx
